Split provider common functions from openstack specifics
